+++ src/models/ProductType.php
@@ -67,6 +67,11 @@ class ProductType extends Model
     public $hasVariants;
 
     /**
+     * @var string Title label
+     */
+    public $variantTitleLabel = 'Title';
+
+    /**
      * @var bool Has variant title field
      */
     public $hasVariantTitleField = true;
@@ -149,8 +154,8 @@ class ProductType extends Model
     {
         return [
             [['id', 'fieldLayoutId', 'variantFieldLayoutId'], 'number', 'integerOnly' => true],
-            [['name', 'handle', 'titleFormat', 'titleLabel'], 'required'],
-            [['name', 'handle', 'descriptionFormat', 'titleLabel'], 'string', 'max' => 255],
+            [['name', 'handle', 'titleFormat', 'titleLabel', 'variantTitleLabel'], 'required'],
+            [['name', 'handle', 'descriptionFormat', 'titleLabel', 'variantTitleLabel'], 'string', 'max' => 255],
             [['handle'], UniqueValidator::class, 'targetClass' => ProductTypeRecord::class, 'targetAttribute' => ['handle'], 'message' => 'Not Unique'],
             [['handle'], HandleValidator::class, 'reservedWords' => ['id', 'dateCreated', 'dateUpdated', 'uid', 'title']],
         ];