+++ src/services/ProductTypes.php
@@ -324,6 +324,7 @@ class ProductTypes extends Component
             'hasVariantTitleField' => $productType->hasVariantTitleField,
             'titleFormat' => $productType->titleFormat,
             'titleLabel' => $productType->titleLabel,
+            'variantTitleLabel' => $productType->variantTitleLabel,
             'skuFormat' => $productType->skuFormat,
             'descriptionFormat' => $productType->descriptionFormat,
             'siteSettings' => []
@@ -414,6 +415,7 @@ class ProductTypes extends Component
             $productTypeRecord->hasVariantTitleField = $data['hasVariantTitleField'];
             $productTypeRecord->titleFormat = $data['titleFormat'] ?: '{product.title}';
             $productTypeRecord->titleLabel = $data['titleLabel'];
+            $productTypeRecord->variantTitleLabel = $data['variantTitleLabel'];
             $productTypeRecord->skuFormat = $data['skuFormat'];
             $productTypeRecord->descriptionFormat = $data['descriptionFormat'];
 
@@ -885,6 +887,7 @@ class ProductTypes extends Component
                 'productTypes.hasDimensions',
                 'productTypes.hasVariants',
                 'productTypes.hasVariantTitleField',
+                'productTypes.variantTitleLabel',
                 'productTypes.titleFormat',
                 'productTypes.titleLabel',
                 'productTypes.skuFormat',